small fix in the code structure

+++ control-flow/match_case_calculator.py
@@ -1,6 +1,6 @@
 # simple cal
-f_num = int(input("Enter the first number:"))
-s_num =int(input("Enter the second number:"))
+f_num = float(input("Enter the first number:"))
+s_num =float(input("Enter the second number:"))
 operator = input("Choose the operation (+, -, *, /):")
 # # operation
 
@@ -20,6 +20,7 @@ operator = input("Choose the operation (+, -, *, /):")
 will never be reached, because the earlier elif operator == "/" runs first — and Python will already try to divide by zero,
 causing a crash (ZeroDivisionError) before that line executes.
 """
+# match_case_calculator.py
 
 match operator:
     case "+":
@@ -29,7 +30,7 @@ match operator:
     case "*":
         print(f"The result is {f_num * s_num}")
     case "/":
-        if s_num ==0:
+        if s_num == 0:
             print("Cannot divide by zero.")
         else:
             print(f"The result is {f_num / s_num}")
